fix(integration tests): remove circular dependency in BerH finanzielleAngaben guards
