+++ app/flows/beratungshilfeFormular/finanzielleAngaben/BeratungshilfeFinanzielleAngabenGuardType.ts
@@ -1,0 +1,5 @@
+import type { GenericGuard } from "~/flows/guards.server";
+import type { BeratungshilfeFinanzielleAngaben } from "./context";
+
+export type BeratungshilfeFinanzielleAngabenGuard =
+  GenericGuard<BeratungshilfeFinanzielleAngaben>;
