+++ app/flows/beratungshilfeFormular/finanzielleAngaben/hasAnyEigentumExceptBankaccountGuard.ts
@@ -1,0 +1,8 @@
+import type { BeratungshilfeFinanzielleAngabenGuard } from "./BeratungshilfeFinanzielleAngabenGuardType";
+
+export const hasAnyEigentumExceptBankaccount: BeratungshilfeFinanzielleAngabenGuard =
+  ({ context }) =>
+    context.hasGeldanlage == "yes" ||
+    context.hasWertsache == "yes" ||
+    context.hasGrundeigentum == "yes" ||
+    context.hasKraftfahrzeug == "yes";
